Refactor/2kotlin (#227)
* Refactor PublicPaperFilterTest
* Turn ScipamatoFilter from an abstract class to an interface
* PublicPaperFilter as data class
* Convert public-entity to kotlin
* Convert public-persistence-api to kotlin
* Convert public-persistence-jooq to kotlin
* Convert public-web to kotlin
* Don't use the audit fields n public entities
* Fix common-entity:test

+++ public/public-persistence-jooq/src/main/java/ch/difty/scipamato/publ/persistence/paper/PublicPaperFilterConditionMapper.kt
@@ -53,7 +53,7 @@ class PublicPaperFilterConditionMapper : AbstractFilterConditionMapper<PublicPap
 
     private fun hasNoOrIdenticalPubYearUntil(filter: PublicPaperFilter) =
         filter.publicationYearUntil == null ||
-            filter.publicationYearFrom.toInt() == filter.publicationYearUntil.toInt()
+            filter.publicationYearFrom == filter.publicationYearUntil
 
     /*
      * Currently does not allow to mix quoted and unquoted terms. If this will
